use Py_ssize_t for width; avoids potential issues with overflow

+++ src/formatters.c
@@ -34,7 +34,7 @@ typedef struct {
     PyObject *encoding;
     int autoline;
     int wrap;
-    long width;
+    Py_ssize_t width;
 
     PyObject *raw_stream;
     int pos;
